用户、资源修改

查看合并请求 lvdeyang/tetris!104

+++ tetris-project/user/cloud/tetris-user/src/main/java/com/sumavision/tetris/user/UserService.java
@@ -14,6 +14,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.alibaba.fastjson.JSON;
+import com.netflix.infix.lang.infix.antlr.EventFilterParser.null_predicate_return;
 import com.sumavision.tetris.auth.token.TerminalType;
 import com.sumavision.tetris.auth.token.TokenDAO;
 import com.sumavision.tetris.auth.token.TokenPO;
@@ -729,7 +730,7 @@ public class UserService{
 		
 		UserVO result = new UserVO().set(user);
 		List<SystemRoleVO> roles = new ArrayList<SystemRoleVO>();
-		if(resetPermissions){
+		if(resetPermissions != null && resetPermissions){
 			List<UserSystemRolePermissionPO> oldBusinessPermissions = userSystemRolePermissionDao.findByUserIdAndRoleType(user.getId(), SystemRoleType.BUSINESS);
 			if(oldBusinessPermissions!=null && oldBusinessPermissions.size()>0){
 				userSystemRolePermissionDao.deleteInBatch(oldBusinessPermissions);
